fixing 26 "PROBLEMS"

+++ include/ExecuteLine.h
@@ -14,10 +14,10 @@ bool executeCommand(vector <string> &command, WareHouse &warehouse);
 
 int csti(string &s, string parameter);
 
-int digits(int num);
+size_t digits(int num);
 
 bool addCustomerGeneral(WareHouse &warehouse, vector<string> &command, int lineNum, bool config);
 
 void printErrorAtLine(int lineNum);
 
-bool isValid(vector<string> &command, int size, int &lineNum, string &usage, bool config);
+bool isValid(vector<string> &command, size_t size, int &lineNum, string &usage, bool config);
